refactor: api 디렉토리 이동 및 적용
- src/app/api -> src/api
